+++ src/parsers/generators/mod.rs
@@ -1,0 +1,12 @@
+pub mod json;
+pub mod stack;
+
+use derive_more::derive::Display;
+
+#[derive(Debug, PartialEq, Display)]
+pub enum BencodeType {
+    Integer,
+    String,
+    List,
+    Dict,
+}
